// src/serializer/transform/model-types.ts
export interface BoundsModel {
  x: number;
  y: number;
  width: number;
  height: number;
}